fdns/control/server/requirements.txt
Django==1.4.3
MySQL-python==1.2.3c1
dj-scaffold==1.0.6
django-appconf==0.6
django-compressor==1.3
django-debug-toolbar==1.0.1
django-devserver==0.7.0
simplejson==3.3.2
six==1.5.2
sqlparse==0.1.10
wsgiref==0.1.2
APScheduler==2.1.1
paramiko==1.13.0
ecdsa==0.9
djangorestframework==2.3.14
requests==2.4.3
dnspython==1.12.0
django-common-helpers==0.5.1
South==0.7.2
django-cron==0.3.0